requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
gunicorn==21.2.0
aiohttp
async_timeout
beautifulsoup4
pydantic
google-genai